+++ src/useForm.ts
@@ -35,6 +35,7 @@ import isMultipleSelect from './utils/isMultipleSelect';
 import compact from './utils/compact';
 import isNullOrUndefined from './utils/isNullOrUndefined';
 import isRadioOrCheckboxFunction from './utils/isRadioOrCheckbox';
+import isWeb from './utils/isWeb';
 import isHTMLElement from './utils/isHTMLElement';
 import { EVENTS, UNDEFINED, VALIDATION_MODE } from './constants';
 import {
@@ -43,12 +44,11 @@ import {
   UnpackNestedValue,
   FieldName,
   InternalFieldName,
-  FieldValue,
   FieldErrors,
   Field,
   FieldRefs,
   UseFormOptions,
-  ValidationRules,
+  RegisterOptions,
   SubmitHandler,
   FieldElement,
   FormStateProxy,
@@ -68,13 +68,13 @@ import {
   DefaultValues,
   FieldError,
   SetFieldValue,
+  FieldArrayDefaultValues,
+  ResetFieldArrayFunctionRef,
+  UseWatchRenderFunctions,
+  RecordInternalNameSet,
 } from './types';
 
 const isWindowUndefined = typeof window === UNDEFINED;
-const isWeb =
-  typeof document !== UNDEFINED &&
-  !isWindowUndefined &&
-  !isUndefined(window.HTMLElement);
 const isProxyEnabled = isWeb ? 'Proxy' in window : typeof Proxy !== UNDEFINED;
 
 export function useForm<
@@ -85,25 +85,19 @@ export function useForm<
   reValidateMode = VALIDATION_MODE.onChange,
   resolver,
   context,
-  defaultValues = {} as UnpackNestedValue<DeepPartial<TFieldValues>>,
+  defaultValues = {} as DefaultValues<TFieldValues>,
   shouldFocusError = true,
   shouldUnregister = true,
   criteriaMode,
 }: UseFormOptions<TFieldValues, TContext> = {}): UseFormMethods<TFieldValues> {
   const fieldsRef = React.useRef<FieldRefs<TFieldValues>>({});
-  const fieldArrayDefaultValuesRef = React.useRef<
-    Record<InternalFieldName<FieldValues>, unknown[]>
-  >({});
-  const fieldArrayValuesRef = React.useRef<
-    Record<InternalFieldName<FieldValues>, unknown[]>
-  >({});
+  const fieldArrayDefaultValuesRef = React.useRef<FieldArrayDefaultValues>({});
+  const fieldArrayValuesRef = React.useRef<FieldArrayDefaultValues>({});
   const watchFieldsRef = React.useRef<InternalNameSet<TFieldValues>>(new Set());
-  const useWatchFieldsRef = React.useRef<
-    Record<string, InternalNameSet<TFieldValues>>
-  >({});
-  const useWatchRenderFunctionsRef = React.useRef<Record<string, () => void>>(
+  const useWatchFieldsRef = React.useRef<RecordInternalNameSet<TFieldValues>>(
     {},
   );
+  const useWatchRenderFunctionsRef = React.useRef<UseWatchRenderFunctions>({});
   const fieldsWithValidationRef = React.useRef<
     FieldNamesMarkedBoolean<TFieldValues>
   >({});
@@ -121,10 +115,7 @@ export function useForm<
   const handleChangeRef = React.useRef<HandleChange>();
   const shallowFieldsStateRef = React.useRef({});
   const resetFieldArrayFunctionRef = React.useRef<
-    Record<
-      InternalFieldName<FieldValues>,
-      (data?: UnpackNestedValue<DeepPartial<TFieldValues>>) => void
-    >
+    ResetFieldArrayFunctionRef<TFieldValues>
   >({});
   const contextRef = React.useRef(context);
   const resolverRef = React.useRef(resolver);
@@ -165,7 +156,7 @@ export function useForm<
   shallowFieldsStateRef.current = shouldUnregister
     ? {}
     : isEmptyObject(shallowFieldsStateRef.current)
-    ? cloneObject(defaultValues, isWeb)
+    ? cloneObject(defaultValues)
     : shallowFieldsStateRef.current;
 
   const updateFormState = React.useCallback(
@@ -234,16 +225,7 @@ export function useForm<
   );
 
   const setFieldValue = React.useCallback(
-    (
-      name: string,
-      rawValue:
-        | FieldValue<TFieldValues>
-        | UnpackNestedValue<DeepPartial<TFieldValues>>
-        | string[]
-        | undefined
-        | null
-        | boolean,
-    ) => {
+    (name: FieldName<TFieldValues>, rawValue: SetFieldValue<TFieldValues>) => {
       const { ref, options } = fieldsRef.current[name] as Field;
       const value =
         isWeb && isHTMLElement(ref) && isNullOrUndefined(rawValue)
@@ -306,11 +288,9 @@ export function useForm<
     (
       name: InternalFieldName<TFieldValues>,
       shouldRender = true,
-    ): Partial<{
-      dirtyFields: FieldNamesMarkedBoolean<TFieldValues>;
-      isDirty: boolean;
-      touched: FieldNamesMarkedBoolean<TFieldValues>;
-    }> => {
+    ): Partial<
+      Pick<FormState<TFieldValues>, 'dirtyFields' | 'isDirty' | 'touched'>
+    > => {
       if (
         readFormStateRef.current.isDirty ||
         readFormStateRef.current.dirtyFields
@@ -337,11 +317,7 @@ export function useForm<
           (readFormStateRef.current.dirtyFields &&
             isDirtyFieldExist !== get(formStateRef.current.dirtyFields, name));
 
-        if (isChanged && shouldRender) {
-          updateFormState({
-            ...state,
-          });
-        }
+        isChanged && shouldRender && updateFormState(state);
 
         return isChanged ? state : {};
       }
@@ -452,9 +428,9 @@ export function useForm<
   );
 
   const setInternalValues = React.useCallback(
-    <T extends FieldName<TFieldValues>, U extends object | unknown[]>(
-      name: T,
-      value: U,
+    (
+      name: FieldName<TFieldValues>,
+      value: SetFieldValue<TFieldValues>,
       { shouldDirty, shouldValidate }: SetValueConfig,
     ) => {
       const data = {};
@@ -487,9 +463,7 @@ export function useForm<
         setInternalValues(name, value, config);
 
         if (fieldArrayNamesRef.current.has(name)) {
-          fieldArrayDefaultValuesRef.current[
-            name
-          ] = value as TFieldValues[string];
+          fieldArrayDefaultValuesRef.current[name] = value;
           resetFieldArrayFunctionRef.current[name]({
             [name]: value,
           } as UnpackNestedValue<DeepPartial<TFieldValues>>);
@@ -524,7 +498,7 @@ export function useForm<
     [updateAndGetDirtyState, setFieldValue, setInternalValues],
   );
 
-  const isFieldWatched = <T extends FieldName<FieldValues>>(name: T) =>
+  const isFieldWatched = <T extends FieldName<TFieldValues>>(name: T) =>
     isWatchAllRef.current ||
     watchFieldsRef.current.has(name) ||
     watchFieldsRef.current.has((name.match(/\w+/) || [])[0]);
@@ -579,7 +553,9 @@ export function useForm<
             ...modeRef.current,
           });
           let state = updateAndGetDirtyState(name, false);
-          let shouldRender = !isEmptyObject(state) || isFieldWatched(name);
+          let shouldRender =
+            !isEmptyObject(state) ||
+            isFieldWatched(name as FieldName<TFieldValues>);
 
           if (
             isBlurEvent &&
@@ -660,7 +636,7 @@ export function useForm<
 
   function setFieldArrayDefaultValues<T extends FieldValues>(data: T): T {
     if (!shouldUnregister) {
-      let copy = cloneObject(data, isWeb);
+      let copy = cloneObject(data);
 
       for (const value of fieldArrayNamesRef.current) {
         if (isKey(value) && !copy[value]) {
@@ -703,7 +679,7 @@ export function useForm<
     return setFieldArrayDefaultValues(
       getFieldsValues(
         fieldsRef,
-        cloneObject(shallowFieldsStateRef.current, isWeb),
+        cloneObject(shallowFieldsStateRef.current),
         shouldUnregister,
       ),
     );
@@ -745,7 +721,7 @@ export function useForm<
   const updateWatchedValue = React.useCallback((name: string) => {
     if (isWatchAllRef.current) {
       updateFormState();
-    } else if (watchFieldsRef) {
+    } else {
       for (const watchField of watchFieldsRef.current) {
         if (watchField.startsWith(name)) {
           updateFormState();
@@ -818,12 +794,9 @@ export function useForm<
       const watchFields = watchId
         ? useWatchFieldsRef.current[watchId]
         : watchFieldsRef.current;
-      const combinedDefaultValues = isUndefined(defaultValue)
-        ? defaultValuesRef.current
-        : defaultValue;
       let fieldValues = getFieldsValues<TFieldValues>(
         fieldsRef,
-        cloneObject(shallowFieldsStateRef.current, isWeb),
+        cloneObject(shallowFieldsStateRef.current),
         shouldUnregister,
         false,
         fieldNames,
@@ -848,12 +821,16 @@ export function useForm<
           fieldValues,
           fieldNames,
           watchFields,
-          isUndefined(defaultValue)
-            ? get(combinedDefaultValues, fieldNames)
-            : (defaultValue as UnpackNestedValue<DeepPartial<TFieldValues>>),
+          isUndefined(get(defaultValuesRef.current, fieldNames))
+            ? defaultValue
+            : get(defaultValuesRef.current, fieldNames),
           true,
         );
       }
+
+      const combinedDefaultValues = isUndefined(defaultValue)
+        ? defaultValuesRef.current
+        : defaultValue;
 
       if (Array.isArray(fieldNames)) {
         return fieldNames.reduce(
@@ -917,7 +894,7 @@ export function useForm<
 
   function registerFieldRef<TFieldElement extends FieldElement<TFieldValues>>(
     ref: TFieldElement & Ref,
-    validateOptions: ValidationRules | null = {},
+    options: RegisterOptions | null = {},
   ): ((name: InternalFieldName<TFieldValues>) => void) | void {
     if (process.env.NODE_ENV !== 'production') {
       if (!ref.name) {
@@ -947,7 +924,7 @@ export function useForm<
     const { name, type, value } = ref;
     const fieldRefAndValidationOptions = {
       ref,
-      ...validateOptions,
+      ...options,
     };
     const fields = fieldsRef.current;
     const isRadioOrCheckbox = isRadioOrCheckboxFunction(ref);
@@ -969,7 +946,7 @@ export function useForm<
     ) {
       fields[name] = {
         ...field,
-        ...validateOptions,
+        ...options,
       };
       return;
     }
@@ -984,7 +961,7 @@ export function useForm<
               } as RadioOrCheckboxOption,
             ],
             ref: { type, name },
-            ...validateOptions,
+            ...options,
           }
         : {
             ...fieldRefAndValidationOptions,
@@ -1009,11 +986,11 @@ export function useForm<
       isEmptyDefaultValue = isUndefined(defaultValue);
 
       if (!isEmptyDefaultValue && !isFieldArray) {
-        setFieldValue(name, defaultValue);
+        setFieldValue(name as FieldName<TFieldValues>, defaultValue);
       }
     }
 
-    if (!isEmptyObject(validateOptions)) {
+    if (!isEmptyObject(options)) {
       set(fieldsWithValidationRef.current, name, true);
 
       if (!isOnSubmit && readFormStateRef.current.isValid) {
@@ -1029,9 +1006,7 @@ export function useForm<
             ? set(validFieldsRef.current, name, true)
             : unset(validFieldsRef.current, name);
 
-          if (previousFormIsValid !== isEmptyObject(error)) {
-            updateFormState();
-          }
+          previousFormIsValid !== isEmptyObject(error) && updateFormState();
         });
       }
     }
@@ -1065,35 +1040,35 @@ export function useForm<
   }
 
   function register<TFieldElement extends FieldElement<TFieldValues>>(
-    rules?: ValidationRules,
+    options?: RegisterOptions,
   ): (ref: (TFieldElement & Ref) | null) => void;
   function register(
     name: FieldName<TFieldValues>,
-    rules?: ValidationRules,
+    options?: RegisterOptions,
   ): void;
   function register<TFieldElement extends FieldElement<TFieldValues>>(
     ref: (TFieldElement & Ref) | null,
-    rules?: ValidationRules,
+    options?: RegisterOptions,
   ): void;
   function register<TFieldElement extends FieldElement<TFieldValues>>(
-    refOrValidationOptions?:
+    refOrRegisterOptions?:
       | FieldName<TFieldValues>
-      | ValidationRules
+      | RegisterOptions
       | (TFieldElement & Ref)
       | null,
-    rules?: ValidationRules,
+    options?: RegisterOptions,
   ): ((ref: (TFieldElement & Ref) | null) => void) | void {
     if (!isWindowUndefined) {
-      if (isString(refOrValidationOptions)) {
-        registerFieldRef({ name: refOrValidationOptions }, rules);
+      if (isString(refOrRegisterOptions)) {
+        registerFieldRef({ name: refOrRegisterOptions }, options);
       } else if (
-        isObject(refOrValidationOptions) &&
-        'name' in refOrValidationOptions
+        isObject(refOrRegisterOptions) &&
+        'name' in refOrRegisterOptions
       ) {
-        registerFieldRef(refOrValidationOptions, rules);
+        registerFieldRef(refOrRegisterOptions, options);
       } else {
         return (ref: (TFieldElement & Ref) | null) =>
-          ref && registerFieldRef(ref, refOrValidationOptions);
+          ref && registerFieldRef(ref, refOrRegisterOptions);
       }
     }
   }
@@ -1111,7 +1086,7 @@ export function useForm<
       let fieldValues = setFieldArrayDefaultValues(
         getFieldsValues(
           fieldsRef,
-          cloneObject(shallowFieldsStateRef.current, isWeb),
+          cloneObject(shallowFieldsStateRef.current),
           shouldUnregister,
           true,
         ),
@@ -1134,9 +1109,7 @@ export function useForm<
         } else {
           for (const field of Object.values(fieldsRef.current)) {
             if (field) {
-              const {
-                ref: { name },
-              } = field;
+              const { name } = field.ref;
 
               const fieldError = await validateField(
                 fieldsRef,
@@ -1222,7 +1195,7 @@ export function useForm<
   };
 
   const reset = (
-    values?: UnpackNestedValue<DeepPartial<TFieldValues>>,
+    values?: DefaultValues<TFieldValues>,
     omitResetState: OmitResetState = {},
   ): void => {
     if (isWeb) {
@@ -1245,10 +1218,7 @@ export function useForm<
     }
 
     fieldsRef.current = {};
-    defaultValuesRef.current = cloneObject(
-      values || defaultValuesRef.current,
-      isWeb,
-    );
+    defaultValuesRef.current = cloneObject(values || defaultValuesRef.current);
     values && renderWatchedInputs('');
 
     Object.values(resetFieldArrayFunctionRef.current).forEach(
@@ -1257,7 +1227,7 @@ export function useForm<
 
     shallowFieldsStateRef.current = shouldUnregister
       ? {}
-      : cloneObject(values, isWeb) || {};
+      : cloneObject(values) || {};
 
     resetRefs(omitResetState);
   };
@@ -1272,13 +1242,13 @@ export function useForm<
 
   React.useEffect(
     () => () => {
-      isUnMount.current = true;
       observerRef.current && observerRef.current.disconnect();
 
       if (process.env.NODE_ENV !== 'production') {
         return;
       }
 
+      isUnMount.current = true;
       Object.values(fieldsRef.current).forEach((field) =>
         removeFieldEventListenerAndRef(field, true),
       );